Copter, Tools, InertialNav: revert APM2 code removal
Re-add the APM2 support in a way that it builds (in under 256kB) and
doesn't create too many code conflicts with master so that future
merges/rebases are easy.

Reverts the following commits or parts of them:

10ec1b82901d2654db6819e1293f651fa4eb51d3
57273b41b41b332726ab7406a502dc7b4bbc3b47
a4be1fede837aa2fa749989c31d74a271857d152
088a49abbfb446dab7df1c2413ea60b23801af62
472a50f9c24fce87d953cd8b365aa09e48131f38
44619525341c39725e555cf629a6048f0ce46741
6f6847c02569ce63ee793576e5fc76fc13810231
7221070533edb933d63c19d9ad124879500c469f
and possibly more as new commits are made that remove APM stuff.

The goal is of course that both: APM users can benefit improvements
in v3.3 and later versions, and the ArduCopter project can benefit
contributions from APM2 board owners.

To avoid potential future merge conflicts, that gps-glitch,
baro-glitch and the fallback mechanism have not been readded which
means that this version is potentially less safe than a supported
build.