Motors: support I2C output to ESCs.
Add an AP_Actuator_Channel subclass that uses a specific simple
I2C-based protocol to talk to ESCs as a bidirectional alertnative
to PWM.  The I2C registers used are documented here:

https://github.com/balrog-kun/tgy/blob/bb878eb8616e203c37c0423badce8529a1c4f025/tgy.asm#L1358-L1370

The class is named AP_Actuator_I2C_SimonK because many other
I2C-based protocols are possible and some are in use (mikrokopter,
early PX4ESC?).  This particular protocol has seen a year and a half
of usage so it's been tested in the field and has the advantage of
being possible on off-the-shelf $10 ESCs.

ArduCopter/ArduCopter.pde creates instances of the class when configured
to do so but only for channels 1..4 meaning only Tri or Quad.
Optimally it should know which channels are ESCs based on frame
settings and also leave the corresponging PWM outputs available for use
as Aux channels for gimbals and things.